Rename show-viewshed-from-point-in-scene to show-exploratory-viewshed-from-point-in-scene
