+++ samples/show-exploratory-viewshed-from-point-in-scene/README.metadata.json
@@ -1,0 +1,39 @@
+{
+    "category": "Analysis",
+    "description": "Perform an exploratory viewshed analysis from a defined vantage point.",
+    "formal_name": "ShowExploratoryViewshedFromPointInScene",
+    "ignore": false,
+    "images": [
+        "show-exploratory-viewshed-from-point-in-scene.png"
+    ],
+    "keywords": [
+        "3D",
+        "exploratory viewshed",
+        "frustum",
+        "geoview-compose",
+        "scene",
+        "visibility analysis",
+        "AnalysisOverlay",
+        "ArcGISSceneLayer",
+        "ArcGISTiledElevationSource",
+        "ExploratoryLocationViewshed",
+        "ExploratoryViewshed"
+    ],
+    "language": "kotlin",
+    "redirect_from": "TODO",
+    "relevant_apis": [
+        "AnalysisOverlay",
+        "ArcGISSceneLayer",
+        "ArcGISTiledElevationSource",
+        "ExploratoryLocationViewshed",
+        "ExploratoryViewshed"
+    ],
+    "snippets": [
+        "src/main/java/com/esri/arcgismaps/sample/showexploratoryviewshedfrompointinscene/components/SceneViewModel.kt",
+        "src/main/java/com/esri/arcgismaps/sample/showexploratoryviewshedfrompointinscene/MainActivity.kt",
+        "src/main/java/com/esri/arcgismaps/sample/showexploratoryviewshedfrompointinscene/screens/ViewshedSlider.kt",
+        "src/main/java/com/esri/arcgismaps/sample/showexploratoryviewshedfrompointinscene/screens/ViewshedOptionsScreen.kt",
+        "src/main/java/com/esri/arcgismaps/sample/showexploratoryviewshedfrompointinscene/screens/MainScreen.kt"
+    ],
+    "title": "Show exploratory viewshed from point in scene"
+}
