+++ samples/show-exploratory-viewshed-from-point-in-scene/show-exploratory-viewshed-from-point-in-scene.png
