+++ samples/show-exploratory-viewshed-from-point-in-scene/src/main/AndroidManifest.xml
