+++ samples/show-exploratory-viewshed-from-point-in-scene/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="show_viewshed_from_point_in_scene_app_name">Show viewshed from point in scene</string>
+    <string name="show_exploratory_viewshed_from_point_in_scene_app_name">Show exploratory viewshed from point in scene</string>
     <string name="elevation_service">https://scene.arcgis.com/arcgis/rest/services/BREST_DTM_1M/ImageServer</string>
     <string name="buildings_layer">
         https://tiles.arcgis.com/tiles/P3ePLMYs2RVChkJx/arcgis/rest/services/Buildings_Brest/SceneServer/layers/0
