--- samples/show-viewshed-from-point-in-scene/README.metadata.json
@@ -1,39 +1,0 @@
-{
-    "category": "Analysis",
-    "description": "Perform an exploratory viewshed analysis from a defined vantage point.",
-    "formal_name": "ShowViewshedFromPointInScene",
-    "ignore": false,
-    "images": [
-        "show-viewshed-from-point-in-scene.png"
-    ],
-    "keywords": [
-        "3D",
-        "frustum",
-        "geoview-compose",
-        "scene",
-        "viewshed",
-        "visibility analysis",
-        "AnalysisOverlay",
-        "ArcGISSceneLayer",
-        "ArcGISTiledElevationSource",
-        "ExploratoryLocationViewshed",
-        "ExploratoryViewshed"
-    ],
-    "language": "kotlin",
-    "redirect_from": "",
-    "relevant_apis": [
-        "AnalysisOverlay",
-        "ArcGISSceneLayer",
-        "ArcGISTiledElevationSource",
-        "ExploratoryLocationViewshed",
-        "ExploratoryViewshed"
-    ],
-    "snippets": [
-        "src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/components/SceneViewModel.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/MainActivity.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/screens/ViewshedSlider.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/screens/MainScreen.kt",
-        "src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/screens/ViewshedOptionsScreen.kt"
-    ],
-    "title": "Show viewshed from point in scene"
-}
